# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Miqaat API",
    "description": "API documentation for Miqaat management system including Census, Events, Sharafs, and related operations",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base URL"
    }
  ],
...
  "paths": {
...
    "/sharafs/{sharaf_id}/status": {
      "put": {
        "tags": ["Sharafs"],
        "summary": "Update sharaf status",
        "parameters": [
          {
            "name": "sharaf_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Sharaf ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["pending", "bs_approved", "confirmed", "rejected", "cancelled"],
                    "description": "New status"
                  }
                },
                "required": ["status"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Sharafs"],
        "summary": "Update sharaf status",
        "parameters": [
          {
            "name": "sharaf_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Sharaf ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["pending", "bs_approved", "confirmed", "rejected", "cancelled"],
                    "description": "New status"
                  }
                },
                "required": ["status"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "description": "Response data"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "example": "The field is required."
          }
        }
      },
...
    }
  }
}